requirements.txt
ffmpeg-python
tgcrypto
uvloop
pyrogram
pytubefix
av
urlextract
pillow
pymongo
instagrapi==2.0.0